+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.4.1] - Unreleased
+
+### auth0-bitbucket-deploy v2.9.1
+### auth0-gitlab-deploy v2.10.1
+### auth0-visualstudio-deploy v2.8.1
+
+- #### Fixed
+  - Roles management bug
+    
 ## [1.4.0] - 2019-06-04
 
 ### auth0-bitbucket-deploy v2.9.0